edit home and about view

+++ app/Http/Controllers/PostController.php
@@ -84,7 +84,12 @@ class PostController extends Controller
     public function show($course_id, Post $post)
     {   $test = (new HomeController)->note();
         $comments = $post->comments()->with('replies')->get();
-        return view('posts.show', ['post' => $post, 'course' => $course_id, 'comments' => $comments,'test'=>$test]);
+        if($post->course_id == $course_id)
+            return view('posts.show', ['post' => $post, 'course' => $course_id, 'comments' => $comments,'test'=>$test]);
+        else {
+            toastr()->error("you entered a wrong url");
+            return redirect()->route('courses.show', $course_id);
+        }
     }
 
     /**
@@ -95,7 +100,12 @@ class PostController extends Controller
      */
     public function edit(Course $course, Post $post)
     {   $test = (new HomeController)->note();
-        return view('posts.edit', ['course' => $course->id, 'post' => $post,'test'=>$test]);
+        if(auth()->id() == $post->user_id)
+            return view('posts.edit', ['course' => $course->id, 'post' => $post,'test'=>$test]);
+        else {
+            toastr()->error("you are not authorized to view this page");
+            return redirect()->route('courses.show', $course->id);
+        }
     }
 
     /**
@@ -130,42 +140,40 @@ class PostController extends Controller
     public function destroy($course_id, Post $post)
     {
         if ($post->image) Storage::delete('public/'.$post->image);
-        $post->delete();
+        if (auth()->id() == $post->user_id)
+            $post->delete();
+        else
+            toastr()->error("you are not authorized to view this page");
         return redirect()->route('courses.show', $course_id);
     }
+
     public function ratePost()
-
     {
+        $post = Post::find(request()->id);
+        $rating = $post->ratings()->where('user_id', auth()->user()->id)->first();
+        if(is_null($rating) ){
+            $rating = new \willvincent\Rateable\Rating;
+            $rating->rating =  request()->rate;
+            $rating->user_id = auth()->user()->id;
+            $post->ratings()->save($rating);
+            return redirect()->back();
+        }
 
-    $post = Post::find(request()->id);
-    $rating = $post->ratings()->where('user_id', auth()->user()->id)->first();
-     if(is_null($rating) ){
+        if( $rating->rating<=5){
 
-        $rating = new \willvincent\Rateable\Rating;
-        $rating->rating =  request()->rate;
-        $rating->user_id = auth()->user()->id;
-        $post->ratings()->save($rating);
-        return redirect()->back();
+            $post->ratings()->where('user_id', auth()->user()->id)->first()->delete($rating);
+            $rating = new \willvincent\Rateable\Rating;
+            $rating->rating =  request()->rate;
+            $rating->user_id = auth()->user()->id;
+            $post->ratings()->save($rating);
+            return redirect()->back();
+        }
+        if( $rating->rating==null){
+
+            return redirect()->back()->with("invalid rating");
+        }
+        else{
+            return redirect()->back()->with("You already made a review");
+        }
     }
-
-    if( $rating->rating<=5){
-
-        $post->ratings()->where('user_id', auth()->user()->id)->first()->delete($rating);
-        $rating = new \willvincent\Rateable\Rating;
-        $rating->rating =  request()->rate;
-        $rating->user_id = auth()->user()->id;
-        $post->ratings()->save($rating);
-        return redirect()->back();
-    }
-    if( $rating->rating==null){
-
-        return redirect()->back()->with("invalid rating");
-    }
-    else{
-        return redirect()->back()->with("You already made a review");
-    }
-
-
-    }
-
 }
